requirements.txt
# Core
tensorflow>=2.10.0
numpy>=1.21.0

# Image Processing
pydicom>=2.3.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# ML Utilities
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Notebooks
jupyter>=1.0.0
tqdm>=4.62.0

# Desktop App (optional)
PyQt5>=5.15.0

# Data
pandas>=1.3.0
